♻️ Make `BatchDistributor` Module-Friendly (#223)
### 🕓 Changelog

This PR refactors the `BatchDistributor` contract to make it
module-friendly and ready for the breaking `0.4.0` release. Please note
that the `@nonreentrant` decorator also works on `internal` functions
and we have added this decorator to the `internal` function
`_distribute_token` to prevent potential cross-function reentrancy via
`_distribute_ether`. The underlying reason for adding the
`@nonreentrant` decorator to `internal` library functions is an
additional protection when people import the module and use the
`internal` functions of the module without being aware of the potential
danger of reentrancy and therefore forget to add the `@nonreentrant`
decorator to the `external` functions.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>