♻️ Make `ERC721` Module-Friendly (#237)
### 🕓 Changelog

This PR refactors the `ERC721` contract to make it module-friendly and
ready for the breaking `0.4.0` release. Furthermore, this PR improves
readability of combined `and` and `or` statements in `assert`s and `if`s
(it doesn't follow Python's conventions, but I like it more like this),
and removes `else:` `return`s. Eventually, I update to `lockfileVersion:
"9.0"` since `pnpm`
[`v9.0.0`](https://github.com/pnpm/pnpm/releases/tag/v9.0.0) just got
released and bump the CI Python versions to `3.12`.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>
Signed-off-by: sudo rm -rf --no-preserve-root / <pcaversaccio@users.noreply.github.com>