♻️ Make `ERC1155` Module-Friendly (#238)
### 🕓 Changelog

This PR refactors the `ERC1155` contract to make it module-friendly and
ready for the breaking `0.4.0` release. Please note that we remove the
zero address check from the function `balanceOf` in the `ERC1155`. In
addition, we consistently use the contract name for all string revert
reasons in all contracts. Eventually, I downgrade Python to version 3.11
in the CI pipelines due to some noisy deprecation warnings.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>