⚡️ Optimise Precompile `modexp` Call in `p256` (#251)
### 🕓 Changelog

Optimise the `raw_call` call to the `modexp` precompile in the `p256`
contract by using `constant` values (i.e. removing the memory variables
`c` and `modexp`) and reordering of function arguments to match the
`_abi_encode` sequence. Furthermore, I bump the `prettier` dependency
to version `3.3.1` and update the `solady` submodule.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>
Co-authored-by: Charles Cooper <cooper.charles.m@gmail.com>