🔒 Amend `p256` Test Coverage (#256)
### 🕓 Changelog

In the test function `testVerifyWycheproofData` for the `p256` contract,
we flip the `s` parameter if it is higher than the malleability
threshold to validate the correctness of the underlying implementation
(h/t @cairoeth for raising this point). We also `bound` the fuzzed
private key in the `testFuzzVerifyWithValidSignature` test to account
for the secp256r1 curve specifications. Finally, we refactor (i.e.
remove) the `_SIGNATURE_INCREMENT` constant in the `ecdsa` contract by
using the built-in Vyper constant `max_value(int256)` instead. You can
validate the equivalence of the two constants by using
[`titanoboa`](https://github.com/vyperlang/titanoboa):

```console
~$ pip install git+https://github.com/vyperlang/titanoboa@vyper-0.4.0
~$ python
>>> import boa
>>> boa.eval("convert(max_value(int256), uint256)") == 57_896_044_618_658_097_711_785_492_504_343_953_926_634_992_332_820_282_019_728_792_003_956_564_819_967
True
```

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>