🔒 Revoke Admin Role From Deployer (#259)
### 🕓 Changelog

This PR revokes the `DEFAULT_ADMIN_ROLE` role from the deployer account
in the `timelock_controller` contract. The underlying reason for this
design is that deployer accounts may forget to revoke the admin rights
from the timelock controller contract after deployment. For further
insights also, see the following issue:
https://github.com/OpenZeppelin/openzeppelin-contracts/issues/3720.

In addition, we remove the redundant ownership transfers in the `erc20`,
`erc721`, and `erc1155` constructors. At initialisation time, the
`owner` role will be already assigned to the `msg.sender` since we
`uses` the `ownable` module.

--------
Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>