♻️ Explicitly Set `nonreentrancy off` Pragma (#320)
### 🕓 Changelog

In line with 🐍 snekmate's focus on explicitness, we explicitly set `#
pragma nonreentrancy off` in all 🐍Vyper files (module and mock
contracts), even though the default is implicit, as shipped in Vyper
commit
[`59baa1b`](https://github.com/vyperlang/vyper/commit/59baa1b7e6bf19214f2d2ecfb68aa243df1e66af).
This change ensures clarity and consistency. The pragma is excluded from
interface files (`.vyi`), where it is disallowed. Furthermore, we bump
`prettier-plugin-solidity` to
[`v2.0.0`](https://github.com/prettier-solidity/prettier-plugin-solidity/releases/tag/v2.0.0)
and enforce `120` line width for Solidity files.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>