🐛 Fix `base64` `_decode` Double Padding Detection (#345)
### 🕓 Changelog

This PR fixes incorrect padding detection in the `base64` `_decode`
function. The logic incorrectly identified `/` or `_` (index `63`) as
the padding character instead of `=` (index `64`), and checked single
padding before double padding. This caused
[Base64](https://en.wikipedia.org/wiki/Base64) strings with `/` or `_`
in the third position (e.g. `AA/A` or `AA_A`) to be incorrectly
truncated, and strings with double padding (`==`) to be mishandled. This
patch corrects the index to `64`, reorders the conditional checks to
handle double padding before single padding, and adds an assertion to
reject invalid padding combinations. Since the `base64` contract is not
used in production, this bug poses _no_ security risk. Thanks to
@Leminkay for identifying this bug!

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>